Bug/issue1759 (#1769)
* fixed broken linking #1759

* Fixed EMF_COMMENT type being written wrong for EMF+ records

* removed a test test

---------

+++ src/EPPlus/Drawing/EMF/Records/PlusRecords/EmfPlusRecord.cs
@@ -40,7 +40,7 @@ namespace OfficeOpenXml.Drawing.EMF.Records
 
         internal virtual void WriteBytes(BinaryWriter bw)
         {
-            bw.Write((uint)Type);
+            bw.Write((ushort)Type);
             bw.Write(PlusFlags);
             bw.Write(Size);
             bw.Write(DataSize);
